requirements.txt
streamlit
pandas
requests
razorpay>=1.4.0
setuptools
flask
pytest
streamlit-autorefresh
reportlab
pillow
plotly
streamlit_autorefresh
fpdf
openai
matplotlib
streamlit-camera-input-live
pytesseract
fitz
frontend
tools
PyPDF2
pytesseract
folium